requirements.txt
rich
omegaconf
scikit-learn
numpy
torchmetrics
torchvision
matplotlib
requests
tqdm
python-dotenv
hydra-core